Fixing errors with OpenMP parallelization and inconsistent values initialized by 'srand(seed)'
OpenMP was parallelizing the code in such a way that randomized doubles were being stored in
random indices of matrices A and B. The trick was to change 'shared' keyword in the '#pragma'
statement to 'private'. Also, we should make 'seed' an unsigned integer because the 'srand'
function takes in an unsigned integer. Thus, if you pass in any standard 'int' variable, the
'rand()' function will not consistently output the same vars across runs.