Adding prefetching code and undoing previous change to OpenMP pragmas. Also updating Makefile.
The prefetching code gives us a 20% speedup, provided you're on an X86 platform. I updated the
Makefile to check whether you're on an X86 platform or not, and if you are, then we call the
prefetching code. I also added a condition 'USE_OPENMP' to define areas of the code file that
only work with OpenMP. Finally, I fixed the incorrect standard deviation calculation because it
was calling OpenMP to get the number of threads, and then it was basing its calculation on that
number. However, not every part of the code uses OpenMP, and we don't need the number of threads
in our calculation anyway.